Some refactor (added interfaces files).
