+++ repository/user.go
@@ -12,28 +12,6 @@ import (
 	"github.com/jackc/pgx/v5/pgtype"
 )
 
-type ErrorFunc func(args ...interface{}) error
-
-var NothingInTableError ErrorFunc = func(args ...interface{}) error {
-	return fmt.Errorf("no rows in result set")
-}
-var InvalidDataInTableError ErrorFunc = func(args ...interface{}) error {
-	return fmt.Errorf("unable to scan: invalid data in table") // ToDo: paste err in message
-}
-var DuplicatedDataError ErrorFunc = func(args ...interface{}) error {
-	return fmt.Errorf("duplicated data in table: %v", args)
-}
-var ConstraintError ErrorFunc = func(args ...interface{}) error {
-	return fmt.Errorf("constraint error: %v", args)
-}
-
-type UserRepositoryInterface interface {
-	Create(ctx context.Context, userInfo models.UserInfo) (int, error)
-	GetById(ctx context.Context, id int) (models.UserInfo, error)
-	GetByUsername(ctx context.Context, username string) (models.UserInfo, error)
-	GetByEmail(ctx context.Context, email string) (models.UserInfo, error)
-}
-
 type PostgresUser struct {
 	db *pgx.Conn
 }
